mass update i aint got shit done

+++ frontend/src/components/AdminBooks.jsx
@@ -3,35 +3,50 @@ import { useNavigate } from "react-router-dom";
 
 function AdminBooks(props) {
     const [books, setBooks] = useState([])
+    const [categories, setCategories] = useState([])
 
     const navigate = useNavigate()
     const addBook = () => {
         navigate('/book-add')
     }
-    const editBook = (bookcode) => {
-        navigate('/book-update/' + bookcode)
+    const editBook = (bookId) => {
+        navigate('/book-update/' + bookId)
     }
 
-    const deleteBook = (bookcode) => {
-        fetch('http://localhost:8080/api/delete/' + bookcode, {
-            method: 'DELETE'
-        })
-            .then((response) => response.json())
-            .then((data) => {
-                setBooks(books.filter(book => book.bookcode !== bookcode));
-                console.log(data)
+    const deleteBook = (bookId) => {
+        if (window.confirm('Delete this item?')) {
+            fetch('http://localhost:9091/server/delete/' + bookId, {
+                method: 'DELETE'
             })
-            .catch((err) => {
-                console.log(err)
-                alert('error')
-            })
+                .then((response) => response.json())
+                .then((data) => {
+                    setBooks(books.filter(book => book.bookId !== bookId));
+                    console.log(data)
+                })
+                .catch((err) => {
+                    console.log(err)
+                    alert(err)
+                })
+        }
     }
     useEffect(() => {
-        fetch('http://localhost:8080/server/books')
+        fetch('http://localhost:9091/server/categories')
             .then((response) => response.json())
-            // .then((data) => console.log(data))
-            .then((data) => setBooks(data))
+            .then((data) => {
+                setCategories(data)
+                //console.log(data)
+                fetch('http://localhost:9091/server/books')
+                    .then((response) => response.json())
+                    // .then((data) => console.log(data))
+                    .then((data) => {
+                        setBooks(data)
+                        //console.log(data)
+                    })
+                    .catch((err) => console.log(err))
+            })
             .catch((err) => console.log(err))
+
+
     }, [])
 
     return (
@@ -49,11 +64,12 @@ function AdminBooks(props) {
                 <table className="table table-striped table-bordered mb-0">
                     <thead>
                         <tr>
-                            <th>BookCode</th>
                             <th>Title</th>
                             <th>Author</th>
+                            <th>Category</th>
                             <th>Release Date</th>
-                            <th>Price</th>
+                            <th>Number of pages</th>
+                            <th>Copies sold</th>
                             <th>Action</th>
                         </tr>
                     </thead>
@@ -61,17 +77,18 @@ function AdminBooks(props) {
                     <tbody>
                         {books.map((book) => (
                             <tr key={book.bookId}>
-                                <td> {book.bookId} </td>
                                 <td> {book.title} </td>
                                 <td> {book.author} </td>
+                                <td> {categories[book.categoryId - 1].categoryName}</td>
                                 <td> {book.releaseDate} </td>
-                                <td> {book.price} </td>
+                                <td> {book.numPage} </td>
+                                <td> {book.numSold}</td>
                                 <td>
                                     <button className="btn btn-primary" onClick={() => {
-                                        editBook(book.bookcode)
+                                        editBook(book.bookId)
                                     }}>View</button>
                                     <button className="btn btn-danger" onClick={() => {
-                                        deleteBook(book.bookcode)
+                                        deleteBook(book.bookId)
                                     }}>Delete</button>
                                 </td>
                             </tr>
